requirements.txt
aiogram
gspread
oauth2client
APScheduler
python-dotenv
